app/src/main/java/com/picpay/desafio/android/view/UserViewModel.kt
package com.picpay.desafio.android.view

import androidx.lifecycle.LiveData
import androidx.lifecycle.MutableLiveData
import androidx.lifecycle.ViewModel
import com.picpay.desafio.android.User
import com.picpay.desafio.android.data.UserRepository

class UserViewModel(private val repository: UserRepository) : ViewModel() {

    private val _users = MutableLiveData<List<User>>()
    val users: LiveData<List<User>>
        get() = _users

    fun fetch() {
        _users.value = repository.getAll()
    }

}